requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
langchain-community==0.0.10
chromadb==0.4.22
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
streamlit==1.28.1
streamlit-chat==0.1.1
python-dotenv==1.0.0
tiktoken==0.5.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3 